[feature] integrated login/signup with index
